\name{check.configVar.cometlist}
\alias{check.configVarcometlist}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Check if all variables have a value related to the function comet.list. Comet.list gives the list of correlation between omic features.
}
\description{
Check if all variables have a value related to the function comet.list. Comet.list gives the list of correlation between omic features.
}
\usage{
check.configVar.cometlist(config.var)
}
\arguments{
    \item{config.var}{
    configuration variables
    }
}

\value{
Nothing
}

\author{
Tiphaine Martin
}

\keyword{internal}

