\name{check.configVar}
\alias{check.configVar}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Check if all variables have a value related to the functions comet and comet.web
}
\description{
Check if all variables have a value related to the functions comet and comet.web
}
\usage{
check.configVar(config.var)
}
\arguments{
    \item{config.var}{
    configuration variables
    }
}

\value{
Nothing
}

\author{
Tiphaine Martin
}

\keyword{internal}

