\name{check.format.mydata}
\alias{check.format.mydata}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Check the format of different data
}
\description{
Check the format of different data
}
\usage{
check.format.mydata(gbl.var, option, numfile)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
    \item{gbl.var}{
list of internal variables
}

    \item{option}{
 option that say if the data to check are main or supplementary data
}
    \item{numfile}{
 the order of file to check from the list of main or supplementary data
}
}

\value{
gbl.var updated or error message
}

\author{
Tiphaine Martin
}

% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{internal}

