\name{create.color.bar}
\alias{create.color.bar}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Create color bar of heatmap
}
\description{
Create color legend for the correlation matrix
}
\usage{
create.color.bar(config.var, gbl.var)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
   \item{config.var}{
list of all variables defined in configuration file or via options of
comet function
}
  \item{gbl.var}{
list of internal variables
}
}

\author{
Tiphaine Martin
}

% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ internal }
