\name{create.tracks.web}
\alias{create.tracks.web}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Create tracks for the web page (see cometweb)
}
\description{
Create tracks for the web page (see cometweb)
}
\usage{
create.tracks.web(config.var, gbl.var)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
   \item{config.var}{
list of all variables defined in configuration file or via options of
comet function
}
  \item{gbl.var}{
list of internal variables
}
}

\author{
Tiphaine Martin
}

% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ internal }
\keyword{ dplot }
