\name{clinCNV}
\alias{allIGtrack}
\alias{allIG}
\alias{allimprintedIG}
\alias{allimprintedIGtrack}
\alias{BindingMotifAll}
\alias{BindingMotifMultiple}
\alias{bindMotifsBiomartTrackDouble}
\alias{bindMotifsBiomartTrackAll}
\alias{bindMotifsBiomartTrackMultiple}
\alias{bindMotifsBiomartTrackSingle}
\alias{BindingMotifSingle}
\alias{chipTFtrack}
\alias{chromHMM_RoadMapSingle}
\alias{chromHMM_RoadMapAllE063}
\alias{chromHMM_RoadMapMultiple}
\alias{chromHMM_RoadMapAll}
\alias{chromatinHMMRoadMapAll}
\alias{chromatinHMMRoadMapSingle}
\alias{chromatinHMMRoadMapMultiple}
\alias{chromhmmNoPattern}
\alias{chromhmmPattern}
\alias{chromhmmtrackone}
\alias{clinCNV}
\alias{clinVariant}
\alias{cpgIstrack}
\alias{coreilVariant}
\alias{cosmicVariant}
\alias{cpgIslandtrack}
\alias{DupTrack}
\alias{dnasetrack}
\alias{DNaseI_RoadMapSingle}
\alias{dgfootprints_RoadMapSingle}
\alias{dyaRM}
\alias{dyaRMtrack}
\alias{enhFANTOMtrack}
\alias{eGTexTrackall}
\alias{eGTexTrackSNP}
\alias{eQTLTrackAll}
\alias{eQTLTrackMultiple}
\alias{eQTLTrackSingle}
\alias{enhRM}
\alias{enhRMtrack}
\alias{gadtrack}
\alias{gwastrack}
\alias{geneRtrack}
\alias{geneNameEnsembl}
\alias{genesUcsctrack}
\alias{gctrack}
\alias{genetrack}
\alias{histonalltrack}
\alias{histoneonetrack}
\alias{iscatrack}
\alias{imprintedGenesGTEx}
\alias{interestgenesENSMBLtrack}
\alias{interesttransENSMBLtrack}
\alias{miRNATargetRegionsBiomartTrack}
\alias{matrix_HiC_Rao}
\alias{metQTLTrackSingle}
\alias{metQTLTrackAll}
\alias{metQTLTrackMultiple}
\alias{OtherRegulatoryRegionsTrackSingle}
\alias{OtherRegulatoryRegionsTrackAll}
\alias{otherRegulatoryRegionsTrackAll}
\alias{otherRegulatoryRegionsTrackSingle}
\alias{PancreasIG}
\alias{PancreasIGtrack}
\alias{PancreasimprintedIG}
\alias{PancreasimprintedIGtrack}
\alias{promRMtrack}
\alias{promRMtrackE063}
\alias{psiGTex}
\alias{psiGTexTrackSNP}
\alias{psiGTexTrackall}
\alias{regulationENSEMBLtrack}
\alias{RegulatoryEvidenceBiomartTrackSingle}
\alias{RegulatoryEvidenceBiomartTrackMultiple}
\alias{RegulatoryEvidenceBiomartTrackAll}
\alias{RegulatoryFeaturesBiomartTrackSingle}
\alias{RegulatoryFeaturesBiomartTrackMultiple}
\alias{RegulatoryFeaturesBiomartTrackAll}
\alias{RegulatorySegmentsBiomartTrackSingle}
\alias{RegulatorySegmentsBiomartTrackMultiple}
\alias{RegulatorySegmentsBiomartTrackAll}
\alias{regulatoryEvidenceBiomartTrackAll}
\alias{regulatoryEvidenceBiomartTrackMultiple}
\alias{regulatoryEvidenceBiomartTrackSingle}
\alias{regulatoryFeaturesBiomartTrackAll}
\alias{regulatoryFeaturesBiomartTrackMultiple}
\alias{regulatoryFeaturesBiomartTrackSingle}
\alias{regulatorySegmentsBiomartTrackAll}
\alias{regulatorySegmentsBiomartTrackMultiple}
\alias{regulatorySegmentsBiomartTrackSingle}
\alias{repeatMaskerTrack}
\alias{rmtrack}
\alias{snptrack}
\alias{snpUCSCtrack}
\alias{StomachIG}
\alias{strutrack}
\alias{StomachIGtrack}
\alias{tfbsFANTOMtrack}
\alias{transENSMBLtrack}
\alias{xenogenestrack}
\docType{data}
\title{
Data sets
}

\description{
Some sample data sets used for the illustrative examples and the vignette.
}

\keyword{ datasets }
\keyword{ internal }





