\name{draw.name.tracks.web}
\alias{draw.name.tracks.web}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Display names of tracks for web page(see cometweb)
}
\description{
Display names of tracks for web page(see cometweb)
}
\usage{
draw.name.tracks.web(config.var, gbl.var)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{config.var}{
list of all variables defined in configuration file or via options of
comet function
}
  \item{gbl.var}{
list of internal variables
}
}

\author{
Tiphaine Martin
}

% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ internal }
\keyword{ print }% __ONLY ONE__ keyword per line
