\name{draw.plot.annotation}
\alias{draw.plot.annotation}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Display the annotation track from ENSEMBL and UCSC
}
\description{
Display the annotation track from ENSEMBL and UCSC between the plot of
pvalue and heatmap
}
\usage{
draw.plot.annotation(config.var, gbl.var)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
\item{config.var}{
list of all variables defined in configuration file or via options of
comet function
}
\item{gbl.var}{
list of internal variables
}
}

\value{
Updated the plot with annotation tracks directly.
}
\author{
Tiphaine Martin
}

% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ internal }
\keyword{ print }% __ONLY ONE__ keyword per line
