\name{draw.plot.axis.data}
\alias{draw.plot.axis.data}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Display the axis data of plot of pvalue
}
\description{
Display the axis data of plot of pvalue
}
\usage{
draw.plot.axis.data(top.vp, config.var, gbl.var)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
\item{top.vp}{
The viewport related to top of plot
}
\item{config.var}{
list of all variables defined in configuration file or via options of
comet function
}
\item{gbl.var}{
list of internal variables
}
}

\value{
Updated the plot with the different axis of pvalue plot
}
\author{
Tiphaine Martin
}

% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ internal }
\keyword{ print }% __ONLY ONE__ keyword per line
