\name{draw.plot.cormatrix.plot}
\alias{draw.plot.cormatrix.plot}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Display the correlation plot at the bottom of the grid
}
\description{
Display the correlation plot at the bottom of the grid
}
\usage{
draw.plot.cormatrix.plot(config.var, gbl.var)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
\item{config.var}{
list of all variables defined in configuration file or via options of
comet function
}
\item{gbl.var}{
list of internal variables
}
}

\value{
Return a viewport containing the correlation matrix
}

\author{
Tiphaine Martin
}

% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ internal }
\keyword{ print }% __ONLY ONE__ keyword per line
