\name{draw.plot.grid.mydata.names}
\alias{draw.plot.grid.mydata.names}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Display the name of elements defined in DATA.FILE
}
\description{
Display the name of elements defined in DATA.FILE
}
\usage{
draw.plot.grid.mydata.names(config.var, gbl.var)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
   \item{config.var}{
list of all variables defined in configuration file or via options of
comet function
}
  \item{gbl.var}{
list of internal variables
}
}

\author{
Tiphaine Martin
}

% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ internal }
\keyword{ print }% __ONLY ONE__ keyword per line
