\name{draw.plot.grid.setup}
\alias{draw.plot.grid.setup}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Set up the grid of plot
}
\description{
Set up the grid of plot
}
\usage{
draw.plot.grid.setup(config.var, gbl.var)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
\item{config.var}{
list of all variables defined in configuration file or via options of
comet function
}
\item{gbl.var}{
list of internal variables
}
}
\value{
Return gbl.var updated with different layout created
}
\author{
Tiphaine Martin
}

% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ internal }
\keyword{ print }% __ONLY ONE__ keyword per line
