\name{draw.plot.mydata.ggbio}
\alias{draw.plot.mydata.ggbio}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
plot tracks created by ggbio that you want to visualise
}
\description{
plot tracks created by ggbio that you want to visualise
}
\usage{
draw.plot.mydata.ggbio(config.var, gbl.var, numfile)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
 \item{config.var}{
list of all variables defined in configuration file or via options of
comet function
}
\item{gbl.var}{
list of internal variables
}

\item{numfile}{
the number of file to visualise
}
}
\value{
Return on the plot the annotation tracks created with the functions of ggbio package
}
\author{
Tiphaine Martin
}

% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ internal }

