\name{fix.values.large}
\alias{fix.values.large}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Fix and update the values of supplementary data
}
\description{
Fix and update the values of supplementary data
}
\usage{
fix.values.large(config.var, gbl.var)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
 \item{config.var}{
list of all variables defined in configuration file or via options of
comet function
}
  \item{gbl.var}{
list of internal variables
}
}

\author{
Tiphaine Martin
}

% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ internal }
\keyword{ dplot }% __ONLY ONE__ keyword per line
