\name{printPlot.comet}
\alias{printPlot.comet}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Create the plot on file from coMet function
}
\description{
Create the plot on file from coMet function
}
\usage{
printPlot.comet(config.var, gbl.var)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{config.var}{
list of all variables defined in configuration file or via options of
comet function
}
  \item{gbl.var}{
list of internal variables
}
}

\author{
Tiphaine Martin
}

% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ internal }
\keyword{ dplot }% __ONLY ONE__ keyword per line
