\name{printPlot.comet.web}
\alias{printPlot.comet.web}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Display the plot from cometWeb funciton
}
\description{
Display the plot from cometWeb funciton
}
\usage{
printPlot.comet.web(config.var, gbl.var)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
\item{config.var}{
list of all variables defined in configuration file or via options of
comet function
}
\item{gbl.var}{
list of internal variables
}
}
\value{
Return the plot produced by comet.web function in a file
} 
\author{
Tiphaine Martin
}

% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ internal }
\keyword{ print }% __ONLY ONE__ keyword per line
