\name{read.config}
\alias{read.config}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Extract the values of variables from configuration file
}
\description{
Extract the values of variables from configuration file
}
\usage{
read.config(config.file, config.var)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
\item{config.file}{
Configuration file
}
\item{config.var}{
list of all variables defined in configuration file or via options of
comet function
}
}
\value{
Return config.var updated with the different values of options found in the configuration file
}
\author{
Tiphaine Martin
}

% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ internal }
\keyword{ dplot }% __ONLY ONE__ keyword per line
