\name{read.file.mydata}
\alias{read.file.mydata}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Read the files of main data and extract data
}
\description{
Read the files of main data and extract data
}
\usage{
read.file.mydata(split.mydata.file, config.var, gbl.var, numfile)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
 \item{split.mydata.file}{
 List of files to main data
}
  \item{config.var}{
list of all variables defined in configuration file or via options of
comet function
}
  \item{gbl.var}{
list of internal variables
}
  \item{numfile}{
 The number of file in the list
}
}

\author{
Tiphaine Martin
}

% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ internal }
\keyword{ dplot }% __ONLY ONE__ keyword per line
