\name{set.image.parameters}
\alias{set.image.parameters}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Set up the parameters of image
}
\description{
Set up the parameters of image
}
\usage{
set.image.parameters(config.var, gbl.var)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
\item{config.var}{
list of all variables defined in configuration file or via options of
comet function
}
\item{gbl.var}{
list of internal variables
}
}
\value{
Return a list of parameter for the production of image
}
\author{
Tiphaine Martin
}

% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ internal }
\keyword{ dplot }% __ONLY ONE__ keyword per line
