\name{create.symbol.list.large}
\alias{create.symbol.list.large}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Create a list of symblo for the supplementary data
}
\description{
Create a list of symblo for the supplementary data
}
\usage{
create.symbol.list.large(config.var, large.split.color.list, gbl.var)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
    \item{config.var}{
list of all variables defined in configuration file or via options of
comet function
}
    \item{large.split.color.list}{
list of color
}
    \item{gbl.var}{
list of internal variables
}
}

\value{
return a list of symbole:
\item{large.split.symbol.list}{a list which contains the list of symbole for the extra data},
\item{large.split.fill.list}{a list which contains the list of fill for the extra data}
}

\author{
Tiphaine Martin
}

% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{internal}
