\name{clinCNV}
\alias{clinCNV}
\alias{chromhmmNoPattern}
\alias{chromhmmPattern}
\alias{chromhmmtrackone}
\alias{clinVariant}
\alias{coreilVariant}
\alias{cosmicVariant}
\alias{cpgIstrack}
\alias{dnasetrack}
\alias{gadtrack}
\alias{gwastrack}
\alias{geneRtrack}
\alias{geneNameEnsembl}
\alias{genesUcsctrack}
\alias{gctrack}
\alias{genetrack}
\alias{histonalltrack}
\alias{histoneonetrack}
\alias{iscatrack}
\alias{regulationENSEMBLtrack}
\alias{rmtrack}
\alias{snptrack}
\alias{snpUCSCtrack}
\alias{strutrack}
\alias{transENSMBLtrack}
\alias{xenogenestrack}
\docType{data}
\title{
Data sets
}

\description{
Some sample data sets used for the illustrative examples and the vignette.
}

\keyword{ datasets }
\keyword{ internal }





