\name{draw.legend}
\alias{draw.legend}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Display the legend of the plot
}
\description{
  display the legend of the plot
  }
\usage{
draw.legend(config.var, gbl.var)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
 \item{config.var}{
list of all variables defined in configuration file or via options of
comet function
}
  \item{gbl.var}{
list of internal variables
}
}

\author{
Tiphaine Martin
}

% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ internal }

